Created projectile pooling system

+++ zombiedefensegame/Assets/GameAssets/Scripts/Gun/IShootable.cs
@@ -2,6 +2,6 @@ namespace TheyAreComing
 {
 	public interface IShootable
 	{
-		public void Damage(float amount);
+		public void Damage(int amount);
 	}
 }
